requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
python-dotenv
google-generativeai>=0.8.3
groq
pinecone>=3.0.0
langchain
langchain-community
langchain-google-genai
langchain-groq
pillow
huggingface_hub
httpx
